fix: add safeguard for avoiding integer division on parsing

+++ atividade_03-05-2026/rational.c
@@ -112,7 +112,7 @@ char *toString(Rational number)
 
 double toDouble(Rational number)
 {
-    return number.numerator / number.denominator;
+    return (double)number.numerator / number.denominator;
 }
 
 int lcm(int a, int b)
